+++ src/include/point_bis.hpp
@@ -4,7 +4,7 @@
 #include <map>
 #include "body2d.hpp"
 #include "dof.hpp"
-#include "disk.hpp" // pourquoi ?? le display peut etre (a mettre au clair)
+#include "disk.hpp" 
 #include "polyg.hpp" // ?
 #include "rline.hpp" // ?
 
@@ -13,7 +13,6 @@ class Point{
 
 	private :
 		double x_ , y_ ;
-		map<body2d*,int> particules_;
 	public:
 		Point(){x_ = 0. ; y_ = 0. ;}
 		~Point(){};
